Rename helper.js to utils.js
